requirements.txt
python-dotenv==1.1.0
streamlit==1.44.1
openai==1.76.0